// apps/frontend/src/shared/api/globals.d.ts
export interface Role {
  /**
   * 角色 ID
   */
  id: string;
  /**
   * 角色名
   */
  name: string;
  /**
   * 角色描述
   */
  description: string | null;
  /**
   * code=代码同步(不可改删),instance=管理 API 创建
   */
  source: 'code' | 'instance';
  /**
   * 创建时间(ISO 8601)
   */
  createdAt: string;
  /**
   * 更新时间(ISO 8601)
   */
  updatedAt: string;
}